Version 5: Release 3 - date/time stamp added, min max functionality working

+++ app/controllers/StationCtrl.java
@@ -15,7 +15,7 @@ public class StationCtrl extends Controller
         render("station.html", station);
     }
 
-    public static void addReading(Long id, int code, double temperature, double windSpeed, double windDirection, int pressure)
+    public static void addReading(Long id, int code, double temperature, double windSpeed, double windDirection, double pressure)
     {
         Reading reading = new Reading(id, code, temperature, windSpeed, windDirection, pressure);
         Station station = Station.findById(id);
